Fixing plugin dependency issue
Both plugins depend on io.netty % netty at differing versions. Which
caused a method to be missing for the bintray plugin since apparently
the newer version of netty doesnt have a setHeader funciton. In order to
fix this I downgraded the version of scrooge-sbt-plugin until a
compatible version was found. Sadly this version is no longer an
autoplugin so we need to import the settings explicitly. Oh well.

Also fixed an issue where `version.sbt` set the version to a SNAPSHOT
which is apparnetly not allowed on bintray...